+++ manager/manager_test.go
@@ -1,11 +1,7 @@
 package manager
 
 import (
-	"context"
-	"errors"
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/netdata/sd/manager/config"
 
@@ -13,182 +9,118 @@ import (
 )
 
 func TestNew(t *testing.T) {
-
-}
-
-func TestManager_Run(t *testing.T) {
 	tests := map[string]struct {
-		configs            []config.Config
-		expectedBeforeStop []*mockPipeline
-		expectedAfterStop  []*mockPipeline
+		prov ConfigProvider
 	}{
-		"add pipeline": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-			},
-			expectedBeforeStop: []*mockPipeline{{name: "name", started: true, stopped: false}},
-			expectedAfterStop:  []*mockPipeline{{name: "name", started: true, stopped: true}},
-		},
-		"remove pipeline": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-				{
-					Source: "source",
-				},
-			},
-			expectedBeforeStop: []*mockPipeline{
-				{name: "name", started: true, stopped: true},
-			},
-			expectedAfterStop: []*mockPipeline{
-				{name: "name", started: true, stopped: true},
-			},
-		},
-		"several equal configs": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-			},
-			expectedBeforeStop: []*mockPipeline{
-				{name: "name", started: true, stopped: false},
-			},
-			expectedAfterStop: []*mockPipeline{
-				{name: "name", started: true, stopped: true},
-			},
-		},
-		"restart pipeline (same source, different config)": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "name1"},
-					Source:   "source",
-				},
-				{
-					Pipeline: &config.PipelineConfig{Name: "name2"},
-					Source:   "source",
-				},
-			},
-			expectedBeforeStop: []*mockPipeline{
-				{name: "name1", started: true, stopped: true},
-				{name: "name2", started: true, stopped: false},
-			},
-			expectedAfterStop: []*mockPipeline{
-				{name: "name1", started: true, stopped: true},
-				{name: "name2", started: true, stopped: true},
-			},
-		},
-		"invalid pipeline config": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "invalid"},
-					Source:   "source",
-				},
-			},
-		},
-		"handle invalid config for running pipeline": {
-			configs: []config.Config{
-				{
-					Pipeline: &config.PipelineConfig{Name: "name"},
-					Source:   "source",
-				},
-				{
-					Pipeline: &config.PipelineConfig{Name: "invalid"},
-					Source:   "source",
-				},
-			},
-			expectedBeforeStop: []*mockPipeline{
-				{name: "name", started: true, stopped: false},
-			},
-			expectedAfterStop: []*mockPipeline{
-				{name: "name", started: true, stopped: true},
-			},
-		},
+		"nil provider":   {prov: nil},
+		"valid provider": {prov: &mockProvider{}},
 	}
 
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			provider := &mockProvider{cfgs: test.configs, ch: make(chan []config.Config)}
-			fact := &mockFactory{}
-			mgr := New(provider)
-			mgr.factory = fact
-
-			var wg sync.WaitGroup
-			ctx, cancel := context.WithCancel(context.Background())
-
-			wg.Add(1)
-			go func() { defer wg.Done(); mgr.Run(ctx) }()
-			time.Sleep(time.Second)
-
-			lock.Lock()
-			assert.Equalf(t, test.expectedBeforeStop, fact.created, "before stop")
-			lock.Unlock()
-			cancel()
-			wg.Wait()
-
-			assert.Equalf(t, test.expectedAfterStop, fact.created, "after stop")
+			assert.NotNil(t, New(test.prov))
 		})
 	}
 }
 
-type (
-	mockProvider struct {
-		cfgs []config.Config
-		ch   chan []config.Config
+func TestManager_Run(t *testing.T) {
+	tests := map[string]func() runSim{
+		"add pipeline": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "name"),
+				},
+				expectedBeforeStop: []*mockPipeline{
+					{name: "name", started: true, stopped: false},
+				},
+				expectedAfterStop: []*mockPipeline{
+					{name: "name", started: true, stopped: true},
+				},
+			}
+			return sim
+		},
+		"remove pipeline": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "name"),
+					prepareEmptyConfig("source"),
+				},
+				expectedBeforeStop: []*mockPipeline{
+					{name: "name", started: true, stopped: true},
+				},
+				expectedAfterStop: []*mockPipeline{
+					{name: "name", started: true, stopped: true},
+				},
+			}
+			return sim
+		},
+		"several equal configs": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "name"),
+					prepareConfig("source", "name"),
+					prepareConfig("source", "name"),
+				},
+				expectedBeforeStop: []*mockPipeline{
+					{name: "name", started: true, stopped: false},
+				},
+				expectedAfterStop: []*mockPipeline{
+					{name: "name", started: true, stopped: true},
+				},
+			}
+			return sim
+		},
+		"restart pipeline (same source, different config)": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "name1"),
+					prepareConfig("source", "name2"),
+				},
+				expectedBeforeStop: []*mockPipeline{
+					{name: "name1", started: true, stopped: true},
+					{name: "name2", started: true, stopped: false},
+				},
+				expectedAfterStop: []*mockPipeline{
+					{name: "name1", started: true, stopped: true},
+					{name: "name2", started: true, stopped: true},
+				},
+			}
+			return sim
+		},
+		"invalid pipeline config": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "invalid"),
+				},
+			}
+			return sim
+		},
+		"handle invalid config for running pipeline": func() runSim {
+			sim := runSim{
+				configs: []config.Config{
+					prepareConfig("source", "name"),
+					prepareConfig("source", "invalid"),
+				},
+				expectedBeforeStop: []*mockPipeline{
+					{name: "name", started: true, stopped: false},
+				},
+				expectedAfterStop: []*mockPipeline{
+					{name: "name", started: true, stopped: true},
+				},
+			}
+			return sim
+		},
 	}
-	mockPipeline struct {
-		name    string
-		started bool
-		stopped bool
-	}
-	mockFactory struct {
-		created []*mockPipeline
-	}
-)
 
-func (m mockProvider) Run(ctx context.Context) {
-	select {
-	case <-ctx.Done():
-	case m.ch <- m.cfgs:
+	for name, sim := range tests {
+		t.Run(name, func(t *testing.T) { sim().run(t) })
 	}
-	<-ctx.Done()
 }
 
-func (m mockProvider) Configs() chan []config.Config {
-	return m.ch
+func prepareConfig(source, name string) config.Config {
+	return config.Config{Pipeline: &config.PipelineConfig{Name: name}, Source: source}
 }
 
-var lock = &sync.Mutex{}
-
-func (m *mockPipeline) Run(ctx context.Context) {
-	lock.Lock()
-	m.started = true
-	lock.Unlock()
-	defer func() { lock.Lock(); defer lock.Unlock(); m.stopped = true }()
-	<-ctx.Done()
-}
-
-func (m *mockFactory) create(cfg config.PipelineConfig) (sdPipeline, error) {
-	lock.Lock()
-	defer lock.Unlock()
-
-	if cfg.Name == "invalid" {
-		return nil, errors.New("mock factory error")
-	}
-	p := &mockPipeline{name: cfg.Name}
-	m.created = append(m.created, p)
-	return p, nil
+func prepareEmptyConfig(source string) config.Config {
+	return config.Config{Source: source}
 }
